requirements.txt
APScheduler==3.6.3
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
envparse==0.2.0
firebase-admin==5.0.0
Flask==2.0.1
google-api-core==1.30.0
google-api-python-client==2.9.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.0
google-cloud-firestore==2.1.3
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.12.3
httplib2==0.19.1
httpx==0.17.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
msgpack==1.0.2
notion-client==0.4.0
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-telegram-bot==13.6
pytz==2021.1
requests==2.25.1
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
tornado==6.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.5
Werkzeug==2.0.1
